Bar charts is formed for juice data.
What changes have you made?
I have made bar chart which will allow represent juice data on the basic of consumption of each juice.
*

Why do we need this change?
These changes help in visualising the juice data in the form of consumption of it.
*

What can this change impact?
We can now easily check which is the most consumed juice and which is less consumed juice.
*